stream

+++ Motor/MotProtocol/MotPacket.c
@@ -184,7 +184,7 @@ void MotPacket_VarWriteResp_BuildInnerHeader(MotPacket_VarWriteResp_T * p_respPa
 }
 uint8_t MotPacket_VarWriteResp_BuildHeader(MotPacket_VarWriteResp_T * p_respPacket, uint8_t varsCount)
 {
-    return Packet_BuildHeader((MotPacket_T *)p_respPacket, MOT_PACKET_VAR_WRITE, varsCount * sizeof(uint16_t) + 4U);
+    return Packet_BuildHeader((MotPacket_T *)p_respPacket, MOT_PACKET_VAR_WRITE, varsCount * sizeof(uint8_t) + 4U);
 }
 
 /******************************************************************************/
